+++ o8g/Scripts/debug.py
@@ -51,9 +51,7 @@ def debugScenario():
    tableCards = [
        "d14694b4-484c-4b45-962e-8cbb636d8a9a" # 200 BP
       ,"c7d128ea-a3b9-4b04-b8b2-a61153b5b2e6" # 400 BP
-      ,"55ab2891-c99e-4647-8a9d-b01fbce3009f" # Chizuru
-      ,"067d592e-2ddf-43f5-82cc-25c70d29a996" # God Rugal
-      # ,"24e99a13-cb42-4e16-9900-78dde12e1a4c" # 600 BP
+      ,"24e99a13-cb42-4e16-9900-78dde12e1a4c" # 600 BP
       # ,"e6e46f83-d089-4762-8d8e-2a3252cfc9db" # 800 BP
    ]
    for i, id in enumerate(tableCards):
@@ -67,10 +65,11 @@ def debugScenario():
       waitForAnimation()
       
    handCards = [
-       "8bb477f9-5004-4018-8d5e-73c6a23e8912" # Char 300 BP
-      ,"e910f462-bea9-4262-b168-c7c512eb6511" # Char 500 BP
-      ,"0fdadc92-0864-46cc-a3ff-c20e2af8249c" # Char 700 BP
-      ,"af43872e-e47d-4fe0-9b55-aedd8a0d0fc7" # Char 800 BP
+       "e590d588-84af-4c26-8dbe-e28d5b626747" # Amakusa
+       # "8bb477f9-5004-4018-8d5e-73c6a23e8912" # Char 300 BP
+      # ,"e910f462-bea9-4262-b168-c7c512eb6511" # Char 500 BP
+      # ,"0fdadc92-0864-46cc-a3ff-c20e2af8249c" # Char 700 BP
+      # ,"af43872e-e47d-4fe0-9b55-aedd8a0d0fc7" # Char 800 BP
       # ,"5e2211a0-e52e-4b7b-b03d-f6ecb3660bb0" # Action 0 SP
       # ,"83c33aa8-5981-4352-8107-cbb7e05547ec" # Action -1 SP
       # ,"80692723-3895-435f-bf8f-e94507704af5" # Action -3 SP