+++ o8g/Scripts/action_labels.py
@@ -25,7 +25,7 @@ def aclb_Freeze(args, x = 0, y = 0):
    else:
       return "Freeze"
 
-      
+
 def aclb_Default(args, x = 0, y = 0):
    card = args[0]
    phaseIdx = getCurrentPhase()
@@ -35,15 +35,15 @@ def aclb_Default(args, x = 0, y = 0):
             return "Remove from attack"
          else:
             return "Attack"
-         
+
       if (not me.isActive or tutorial) and phaseIdx == BlockPhase:
          if isBlocking(card):
             return "Remove from combat"
          else:
             return "Counter-attack"
-      
+
       return "Use ability"
-      
+
    elif isButton(card):
       return "Activate"
    elif isAvatar(card):
@@ -51,7 +51,7 @@ def aclb_Default(args, x = 0, y = 0):
    else:
       return "Resolve"
 
-      
+
 def aclb_ChangeSlot(args, x = 0, y = 0):
    card = args[0]
    targets = getTargetedCards(card, True, card.controller == me)
@@ -60,15 +60,15 @@ def aclb_ChangeSlot(args, x = 0, y = 0):
    else:
       return "Change slot"
 
-      
+
 def aclb_Ability(args, x = 0, y = 0):
    card = args[0]
-   if card.alternate == 'noability':
+   if card.alternate == "noability":
       return "Restore abilities"
    else:
       return "Lose abilities"
 
-      
+
 def aclb_Destroy(args, x = 0, y = 0):
    card = args[0]
    if isCharacter(card):
@@ -76,7 +76,7 @@ def aclb_Destroy(args, x = 0, y = 0):
    else:
       return "Discard"
 
-      
+
 def aclb_RevealTopDeck(group, x = 0, y = 0):
    if len(group) > 0:
       if group[0].isFaceUp:
@@ -85,53 +85,58 @@ def aclb_RevealTopDeck(group, x = 0, y = 0):
          return "Reveal top card"
          return "Discard"
 
-      
+
 def aclb_PlayAuto(group, x = 0, y = 0):
-   return aclb_setting('Play automation', 'PlayAuto')
+   return aclb_setting("Play automation", "PlayAuto")
 
-      
+
 def aclb_PhaseAuto(group, x = 0, y = 0):
-   return aclb_setting('Phase auto advance', 'PhaseAuto')
+   return aclb_setting("Phase auto advance", "PhaseAuto")
 
-      
+
 def aclb_ActivateAuto(group, x = 0, y = 0):
-   return aclb_setting(u'Activate {} abilities and effects'.format(InstantUniChar), 'Activate')
+   return aclb_setting(u"Activate {} abilities and effects".format(InstantUniChar), "Activate")
 
-      
+
 def aclb_WinForms(group, x = 0, y = 0):
-   return aclb_setting('Show alert messages', 'WinForms')
+   return aclb_setting("Show alert messages", "WinForms")
 
-      
+
 def aclb_Sounds(group, x = 0, y = 0):
-   return aclb_setting('Game sounds', 'Sounds')
+   return aclb_setting("Game sounds", "Sounds")
 
-      
+
 def aclb_WelcomeScreen(group, x = 0, y = 0):
-   return aclb_setting('Show welcome screen on start', 'WelcomeScreen')
+   return aclb_setting("Show welcome screen on start", "WelcomeScreen")
 
-      
+
 def aclb_NextOrPass(group, x = 0, y = 0):
    if me.isActive:
       return "Next phase"
    else:
       return "Pass"
-      
-      
+
+
 def aclb_unitedAttack(group, x = 0, y = 0):
-   if getGlobalVar('UnitedAttack'):
+   if getGlobalVar("UnitedAttack"):
       return "Join United Attack"
    else:
       return "United Attack"
-      
-      
+
+
 def aclb_RemovedDefault(group, x = 0, y = 0):
-   if me.isActive and getCurrentPhase() == MainPhase and getRule('play_removed'):
-      return 'Play card'
-      
-   
+   if me.isActive and getCurrentPhase() == MainPhase and getRule("play_removed"):
+      return "Play card"
+
+
+def aclb_SetupOrStart(group, x = 0, y = 0):
+   if setupDone:
+      return "Start game"
+
+
 #---------------------------------------------------------------------------
 # Helpers
 #---------------------------------------------------------------------------
 
 def aclb_setting(label, setting):
-   return "{}{}".format(u'\u2714 ' if settings[setting] else u' \u2002 \u2002 ', label)
+   return "{}{}".format(u"\u2714 " if settings[setting] else u" \u2002 \u2002 ", label)