+++ o8g/Scripts/windows.py
@@ -90,9 +90,16 @@ def getGameVersion():
 
 def showChangelog():
    switchSetting("GameVersion", getGameVersion())
-   msg = u"""What's new in version 0.9.11
+   msg = u"""What's new in version 0.10.0
 
-- 
+- Increased the font size of the rule text of the cards for a better legibility.
+- Replaced the font of the BP & SP values.
+- Confirm boxes shows a "Do not show again" check box option.
+- New options in the Welcome screen to load decks.
+- Nerfed Amakusa SP gain per KOed character.
+- Nerfed Kazuki: now it deals 400 damage.
+- Improved game UX.
+- Several bug fixes.
 """.format(TriggerUniChar)
    askChoice(msg, [], [], ["Close"])