+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Increased the font size of the rule text for a better legibility.
 + Confirm form can show a "Do not show again" check box option.
 + Replaced BP & SP font.
++ Nerfed Amakusa SP gain per KOed chararacter.
 
 ## 0.9.10
 + Fixes card images not rendering because some PNG images were compressed at 8 bit depth.