+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Nerfed Amakusa SP gain per KOed chararacter.
 + Updated illustration of the card Michelle Heart.
 + Improved UX.
++ Fixed error when a target of a rule has quantity with no min value.
 
 ## 0.9.10
 + Fixes card images not rendering because some PNG images were compressed at 8 bit depth.