+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 + Bugfix: Error when the target of a rule has a quantity with no min value.
 + Bugfix: `copyAbility()` does not update properly the copied ability for the other players.
 + Bugfix: game events from auto abilities with suffix "once" are deleted if the card is removed from the table (affected Crossover and Grandmaster).
++ Bugfix `Rules.card` can be deleted during execution of the effects of the rule.
 
 ## 0.9.10
 + Fixes card images not rendering because some PNG images were compressed at 8 bit depth.